+++ soft_i2c/soft_i2c.c
@@ -332,9 +332,9 @@ static uint8_t misaka_soft_i2c_bit_send_address(const misaka_soft_i2c_t *ops, mi
  * @param ops i2c设备
  * @param msgs 消息对象
  * @param num 消息数量
- * @return uint16_t @c 操作的消息数
+ * @return uint32_t @c 操作的消息数
  */
-static uint16_t misaka_soft_i2c_bit_xfer(const misaka_soft_i2c_t *ops, misaka_soft_i2c_message *msgs, uint16_t num)
+static uint32_t misaka_soft_i2c_bit_xfer(const misaka_soft_i2c_t *ops, misaka_soft_i2c_message *msgs, uint16_t num)
 {
 	misaka_soft_i2c_message_t msg;
 	uint32_t i, ret;
@@ -399,9 +399,9 @@ static uint16_t misaka_soft_i2c_bit_xfer(const misaka_soft_i2c_t *ops, misaka_so
  * @param num 消息数量
  * @return uint16_t @c 操作的消息数
  */
-uint16_t misaka_soft_i2c_transfer(const misaka_soft_i2c_t *ops, misaka_soft_i2c_message *msgs, uint16_t num)
+uint32_t misaka_soft_i2c_transfer(const misaka_soft_i2c_t *ops, misaka_soft_i2c_message *msgs, uint32_t num)
 {
-	uint16_t ret;
+	uint32_t ret;
 
 	ops->mutex_take();
 	ret = misaka_soft_i2c_bit_xfer(ops, msgs, num);
@@ -421,11 +421,11 @@ uint16_t misaka_soft_i2c_transfer(const misaka_soft_i2c_t *ops, misaka_soft_i2c_
  * @param flags 标志
  * @param buf 待发送数据
  * @param len 发送数据长度
- * @return uint16_t @c 操作的消息数
+ * @return uint32_t @c 操作的消息数
  */
-uint16_t misaka_soft_i2c_master_send(const misaka_soft_i2c_t *ops, uint16_t addr, uint16_t flags, uint8_t *buf, uint32_t len)
+uint32_t misaka_soft_i2c_master_send(const misaka_soft_i2c_t *ops, uint16_t addr, uint16_t flags, uint8_t *buf, uint32_t len)
 {
-	uint16_t ret;
+	uint32_t ret;
 	misaka_soft_i2c_message msg;
 
 	msg.addr = addr;
@@ -445,11 +445,11 @@ uint16_t misaka_soft_i2c_master_send(const misaka_soft_i2c_t *ops, uint16_t addr
  * @param flags 标志
  * @param rxbuf 待接收数据
  * @param rxlen 接收数据长度
- * @return uint16_t @c 操作的消息数
+ * @return uint32_t @c 操作的消息数
  */
-uint16_t misaka_soft_i2c_master_recv(const misaka_soft_i2c_t *ops, uint16_t addr, uint16_t flags, uint8_t *rxbuf, uint32_t rxlen)
+uint32_t misaka_soft_i2c_master_recv(const misaka_soft_i2c_t *ops, uint16_t addr, uint16_t flags, uint8_t *rxbuf, uint32_t rxlen)
 {
-	uint16_t ret;
+	uint32_t ret;
 	misaka_soft_i2c_message msg;
 
 	msg.addr = addr;
@@ -470,11 +470,11 @@ uint16_t misaka_soft_i2c_master_recv(const misaka_soft_i2c_t *ops, uint16_t addr
  * @param txlen 发送数据长度
  * @param rxbuf 待接收数据
  * @param rxlen 接收数据长度
- * @return uint16_t @c 操作的消息数
+ * @return uint32_t @c 操作的消息数
  */
-uint16_t misaka_soft_i2c_master_send_then_recv(const misaka_soft_i2c_t *ops, uint16_t addr, uint8_t *txbuf, uint32_t txlen, uint8_t *rxbuf, uint32_t rxlen)
+uint32_t misaka_soft_i2c_master_send_then_recv(const misaka_soft_i2c_t *ops, uint16_t addr, uint8_t *txbuf, uint32_t txlen, uint8_t *rxbuf, uint32_t rxlen)
 {
-	uint16_t ret;
+	uint32_t ret;
 	misaka_soft_i2c_message msg[2];
 
 	msg[0].addr = addr;
@@ -500,11 +500,11 @@ uint16_t misaka_soft_i2c_master_send_then_recv(const misaka_soft_i2c_t *ops, uin
  * @param txlen1 发送数据长度1
  * @param txbuf2 待发送数据2
  * @param txlen2 发送数据长度2
- * @return uint16_t @c 操作的消息数
+ * @return uint32_t @c 操作的消息数
  */
-uint16_t misaka_soft_i2c_master_send_then_send(const misaka_soft_i2c_t *ops, uint16_t addr, uint8_t *txbuf1, uint32_t txlen1, uint8_t *txbuf2, uint32_t txlen2)
+uint32_t misaka_soft_i2c_master_send_then_send(const misaka_soft_i2c_t *ops, uint16_t addr, uint8_t *txbuf1, uint32_t txlen1, uint8_t *txbuf2, uint32_t txlen2)
 {
-	uint16_t ret;
+	uint32_t ret;
 	misaka_soft_i2c_message msg[2];
 
 	msg[0].addr = addr;
